Share memory reduction logic between WSL2 and WSLC
Extract the memory compaction and cache reclaim thread into
StartMemoryReductionThread() in util.cpp. WSLC now runs the same
idle-based compaction and drop_caches logic that WSL2 uses, improving
page reporting efficiency and reducing host memory pressure.

- Add StartMemoryReductionThread() and MemoryReductionMode to util.h/cpp
- WSL2 main.cpp delegates to shared implementation
- WSLC WSLCInit.cpp calls StartMemoryReductionThread(DropCache)
- Remove duplicated GetUserCpuTime/GetMemoryInUse from main.cpp
