+++ src/linux/init/WSLCInit.cpp
@@ -1052,6 +1052,12 @@ int WSLCEntryPoint(int Argc, char* Argv[])
     }
 
     //
+    // Start memory reduction thread (compaction + drop_caches when idle).
+    //
+
+    UtilStartMemoryReductionThread(MemoryReductionModeDropCache);
+
+    //
     // Establish the message channel with the service via hvsocket.
     //
